Create README.md Toms-Gems was founded by Thomas Keith to create jobs in my area.  I am seeking investors committed to success in development of Toms-Gems.
Toms-Gems founder Thomas Keith is committed in development of a company that will create jobs for people in my area.  I am seeking investors for the development of internet based companies that are committed to success and willing in the further development of Toms-Gems.  Thank you sincerely for your time.